refactor: Refatora o NinjaController para incluir ResponseEntity as rotas e customiza as respostas das requisições para o cliente.

+++ src/main/java/dev/kisuk1/CadastroDeNinjas/Controller/NinjaController.java
@@ -3,6 +3,9 @@ package dev.kisuk1.CadastroDeNinjas.Controller;
 import dev.kisuk1.CadastroDeNinjas.DTO.NinjaDTO;
 import dev.kisuk1.CadastroDeNinjas.Mapper.NinjaMapper;
 import dev.kisuk1.CadastroDeNinjas.Service.NinjaService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,27 +26,48 @@ public class NinjaController {
     }
 
     @PostMapping("/post")
-    public NinjaDTO postNinja(@RequestBody NinjaDTO newNinjaDTO) {
-        return ninjaService.createNinja(newNinjaDTO);
+    public ResponseEntity<String> postNinja(@RequestBody NinjaDTO newNinjaDTO) {
+        NinjaDTO postNinjaResponse = ninjaService.createNinja(newNinjaDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja Criado com Sucesso. Nome: " + postNinjaResponse.getNome());
     }
 
     @GetMapping("/getAll")
-    public List<NinjaDTO> getAllNinja() {
-        return ninjaService.readAllNinja();
+    public ResponseEntity<List<NinjaDTO>> getAllNinja() {
+        List<NinjaDTO> listNinjaResponse = ninjaService.readAllNinja();
+        return ResponseEntity.ok(listNinjaResponse);
     }
 
     @GetMapping("/getById/{id}")
-    public NinjaDTO getByIdNinja(@PathVariable Long id) {
-        return ninjaService.readByIdNinja(id);
+    public ResponseEntity<?> getByIdNinja(@PathVariable Long id) {
+        NinjaDTO ninjaIdResponse = ninjaService.readByIdNinja(id);
+        if (ninjaIdResponse != null) {
+            return ResponseEntity.ok(ninjaIdResponse);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja não existe em nossa Base de Dados.");
+        }
     }
 
     @PutMapping("/put/{id}")
-    public NinjaDTO putNinja(@PathVariable Long id, @RequestBody NinjaDTO upNinjaDTO) {
-        return ninjaService.updateNinja(id, upNinjaDTO);
+    public ResponseEntity<?> putNinja(@PathVariable Long id, @RequestBody NinjaDTO upNinjaDTO) {
+        if (ninjaService.readByIdNinja(id) != null) {
+            NinjaDTO upNinjaResponse = ninjaService.updateNinja(id, upNinjaDTO);
+            return ResponseEntity.ok(upNinjaResponse);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja não existe em nossa Base de Dados.");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteNinja(@PathVariable Long id) {
-        ninjaService.deleteNinja(id);
+    public ResponseEntity<String> deleteNinja(@PathVariable Long id) {
+        if (ninjaService.readByIdNinja(id) != null) {
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.ok("Ninja Deletado com Sucesso.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja não existe em nossa Base de Dados.");
+        }
     }
 }
